pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pirogue_admin_client*"]
namespaces = false


[project]
name = "pirogue-admin-client"
version = "2.0.4"
description = "Administration client for the PiRogue"
readme = "README.md"
license = {text = "GPL-3.0+"}

requires-python = ">= 3.9"
dependencies = [
  "cryptography~=38.0.4",
  "grpcio",
  "pirogue_admin_api==2.0.2",
  "pyyaml"
]

authors = [
  {name = "Christophe Andral", email = "christophe@andral.fr"},
]

maintainers = [
  {name = "U+039b", email = "hello@pts-project.org"}
]

[project.scripts]
pirogue-admin-client = "pirogue_admin_client.cli:main"

[project.urls]
repository = "https://github.com/PiRogueToolSuite/pirogue-admin"
issues = "https://github.com/PiRogueToolSuite/pirogue-admin/issues"